change(web): iconの背景を個別で指定できるように変更
